py_module!: extra_init escape hatch + convert 9 stub modules
Grow `py_module!` with an optional `extra_init: |ns| { ... }` block so
modules that need PyPy's `buildloaders` / `startup` semantics — array-
loop constants, cfg-gated entries, runtime-derived values — can fit
the macro without spilling back into a hand-written register_module.

Convert errno (host_env vs fallback constant tables, unix-only extras),
atexit (callback shells), _contextvars (ContextVar / Token stubs),
_abc (_abc_instancecheck via baseobjspace::isinstance_w),
_codecs (lookup_error / encode / decode identity shells),
unicodedata (normalize / category / lookup stubs),
_posixshmem (entire surface gated on `cfg(all(unix, host_env))`),
gc (collect drives majit_gc::collect_full, DEBUG_* constants),
_opcode (stack_effect / has_* / get_opname stubs).

Each conversion deletes the module's `moduledef.rs` (one-line init)
and `interp_x.rs` (verbatim register_module body), leaving `mod.rs`
as the single declarative source.  importing.rs updated to call
`module::name::init` directly.  Surface and registration order
unchanged; check.py 39/39+39/39+2/2.
