--- pyre/pyre-interpreter/src/module/_abc/moduledef.rs
@@ -1,7 +1,0 @@
-//! _abc module definition — PyPy: pypy/module/_abc/moduledef.py
-
-use crate::DictStorage;
-
-pub fn init(ns: &mut DictStorage) {
-    super::interp_abc::register_module(ns);
-}
