--- pyre/pyre-interpreter/src/module/_codecs/moduledef.rs
@@ -1,7 +1,0 @@
-//! _codecs module definition — PyPy: pypy/module/_codecs/moduledef.py
-
-use crate::DictStorage;
-
-pub fn init(ns: &mut DictStorage) {
-    super::interp_codecs::register_module(ns);
-}
