--- pyre/pyre-interpreter/src/module/_contextvars/moduledef.rs
@@ -1,7 +1,0 @@
-//! _contextvars module definition — PyPy: pypy/module/_contextvars/moduledef.py
-
-use crate::DictStorage;
-
-pub fn init(ns: &mut DictStorage) {
-    super::interp_contextvars::register_module(ns);
-}
