--- pyre/pyre-interpreter/src/module/gc/moduledef.rs
@@ -1,7 +1,0 @@
-//! gc module definition — PyPy: pypy/module/gc/moduledef.py
-
-use crate::DictStorage;
-
-pub fn init(ns: &mut DictStorage) {
-    super::interp_gc::register_module(ns);
-}
